+++ src/DocxMerger/DocxMergerField.cpp
@@ -33,17 +33,17 @@ void DocxMergerField::setClipboardValue(double p_value) {
 } // setClipboardValue( double )
 
 void DocxMergerField::deserialize(UnzipFile* p_unzipFile) {
-  p_unzipFile ->insertPtrBySeq(p_unzipFile ->readNum<uint32>(), this);
+  p_unzipFile ->insertPtrBySeq(p_unzipFile ->readNum<size_t>(), this);
 
-  m_item = (DocxMergerItem*) p_unzipFile ->readNum<uint32>();
+  m_item = (DocxMergerItem*) p_unzipFile ->readNum<size_t>();
   m_name = p_unzipFile ->readStr();
 } // deserialize
 
 void DocxMergerField::link(UnzipFile* p_unzipFile) {
-  const map<uint32, void*>* l_ptrsBySeq = p_unzipFile ->getPtrsBySeq();
-  uint32 l_ptrSeq;
+  const map<size_t, void*>* l_ptrsBySeq = p_unzipFile ->getPtrsBySeq();
+  size_t l_ptrSeq;
 
-  l_ptrSeq = (uint32) m_item;
+  l_ptrSeq = (size_t) m_item;
   m_item = (DocxMergerItem*) l_ptrsBySeq ->find(l_ptrSeq) ->second;
   m_itemFile = m_item ->getItemFile();
   m_file = m_itemFile ->getFile();
